tests/livequestions: adjust tests to new livequestion creating
