+++ tests/livequestions/test_livequestion_api.py
@@ -3,7 +3,71 @@ from django.urls import reverse
 
 from meinberlin.apps.livequestions import models
 from meinberlin.apps.livequestions import phases
+from meinberlin.test.helpers import freeze_phase
+from meinberlin.test.helpers import freeze_post_phase
+from meinberlin.test.helpers import freeze_pre_phase
 from meinberlin.test.helpers import setup_phase
+
+
+@pytest.mark.django_db
+def test_anonymous_can_create_livequestion_during_active_phase(
+        apiclient, category_factory, phase_factory):
+
+    phase = phase_factory(phase_content=phases.IssuePhase())
+    module = phase.module
+    category = category_factory(module=module)
+    url = reverse('questions-list',
+                  kwargs={'module_pk': module.pk})
+    data = {
+        'text': 'I have a question',
+        'category': category.pk
+    }
+    with freeze_phase(phase):
+        response = apiclient.post(url, data, format='json')
+        assert response.status_code == 201
+
+    live_question = models.LiveQuestion.objects.first()
+    assert live_question.text == 'I have a question'
+    assert live_question.category == category
+
+
+@pytest.mark.django_db
+def test_user_cannot_create_livequestion_during_pre_phase(
+        apiclient, user, category_factory, phase_factory):
+
+    phase = phase_factory(phase_content=phases.IssuePhase())
+    module = phase.module
+    category = category_factory(module=module)
+    url = reverse('questions-list',
+                  kwargs={'module_pk': module.pk})
+    data = {
+        'text': 'I have a question',
+        'category': category.pk
+    }
+    assert apiclient.login(username=user.email, password='password')
+    with freeze_pre_phase(phase):
+        response = apiclient.post(url, data, format='json')
+        assert response.status_code == 403
+
+
+@pytest.mark.django_db
+def test_initiator_cannot_create_livequestion_during_post_phase(
+        apiclient, category_factory, phase_factory):
+
+    phase = phase_factory(phase_content=phases.IssuePhase())
+    module = phase.module
+    category = category_factory(module=module)
+    initiator = module.project.organisation.initiators.first()
+    url = reverse('questions-list',
+                  kwargs={'module_pk': module.pk})
+    data = {
+        'text': 'I have a question',
+        'category': category.pk
+    }
+    assert apiclient.login(username=initiator.email, password='password')
+    with freeze_post_phase(phase):
+        response = apiclient.post(url, data, format='json')
+        assert response.status_code == 403
 
 
 @pytest.mark.django_db
